chore: separate zeebe specific utils
