+++ lib/base/VariableResolver.js
@@ -1,6 +1,6 @@
 import { getBusinessObject, is } from 'bpmn-js/lib/util/ModelUtil';
-import CachedValue from '../util/CachedValue';
-import { getParents, getScope } from '../util/scopeUtil';
+import CachedValue from './util/CachedValue';
+import { getParents, getScope } from './util/scopeUtil';
 
 /**
  * @typedef {Object} AdditionalVariable
