+++ lib/base/util/CachedValue.js
