+++ lib/base/util/ExtensionElementsUtil.js
