+++ lib/base/util/scopeUtil.js
