+++ lib/camunda/VariableResolver.js
@@ -1,10 +1,10 @@
 import { getProcessVariables } from '@bpmn-io/extract-process-variables';
-import { BaseVariableResolver } from './base/VariableResolver';
+import { BaseVariableResolver } from '../base/VariableResolver';
 
 /**
  * The Camunda 7 Implementation for the VariableResolver.
  */
-export class CamundaVariableResolver extends BaseVariableResolver {
+export default class CamundaVariableResolver extends BaseVariableResolver {
   constructor(eventBus, bpmnjs) {
     super(eventBus, bpmnjs);
     this._baseExtractor = getProcessVariables;
