+++ lib/index.js
@@ -1,5 +1,5 @@
-import { CamundaVariableResolver } from './CamundaVariableResolver';
-import { ZeebeVariableResolver } from './ZeebeVariableResolver';
+import CamundaVariableResolver from './camunda/VariableResolver';
+import ZeebeVariableResolver from './zeebe/VariableResolver';
 import VariableProvider from './VariableProvider';
 
 export const ZeebeVariableResolverModule = {
