+++ lib/zeebe/VariableResolver.js
@@ -1,5 +1,5 @@
 import { getProcessVariables } from '@bpmn-io/extract-process-variables/zeebe';
-import { BaseVariableResolver } from './base/VariableResolver';
+import { BaseVariableResolver } from '../base/VariableResolver';
 import { parseIoMappings } from './util/feelUtility';
 
 const HIGH_PRIORITY = 2000;
@@ -7,7 +7,7 @@ const HIGH_PRIORITY = 2000;
 /**
  * The Camunda 8 Implementation for the VariableResolver.
  */
-export class ZeebeVariableResolver extends BaseVariableResolver {
+export default class ZeebeVariableResolver extends BaseVariableResolver {
   constructor(eventBus, bpmnjs) {
     super(eventBus, bpmnjs);
     this._baseExtractor = getProcessVariables;
