+++ lib/zeebe/util/VariableContext.js
