+++ lib/zeebe/util/feelUtility.js
@@ -4,8 +4,8 @@ import {
 } from '@lezer/lr';
 
 import { EntriesContext } from './VariableContext';
-import { getExtensionElementsList } from './ExtensionElementsUtil';
-import { getParents } from './scopeUtil';
+import { getExtensionElementsList } from '../../base/util/ExtensionElementsUtil';
+import { getParents } from '../../base/util/scopeUtil';
 
 
 export function parseIoMappings(variables) {
@@ -124,7 +124,7 @@ function resolveReferences(variablesToResolve, allVariables) {
  * @param {Variables} variables
  * @returns {EntriesContext}
  */
-function getResultContext(expression, variables = {}) {
+export function getResultContext(expression, variables = {}) {
   const contextTracker = trackVariables(variables, EntriesContext);
 
   // This is a hack to get the latest variables from the context tracker
@@ -237,7 +237,7 @@ function toOptimizedFormat(variables) {
 /**
  * Transforms EntriesContext to the format required by the feel-editor
  */
-function toUnifiedFormat(variables) {
+export function toUnifiedFormat(variables) {
   if (!variables) {
     return;
   }
